+++ packages/block-editor/src/components/inserter/hooks/use-block-types-state.js
@@ -1,17 +1,13 @@
 import {
-	getBlockType,
 	createBlock,
 	createBlocksFromInnerBlocksTemplate,
 	store as blocksStore,
 	parse,
 } from '@wordpress/blocks';
-import { useSelect, useDispatch } from '@wordpress/data';
+import { useSelect } from '@wordpress/data';
 import { useCallback, useMemo } from '@wordpress/element';
-import { store as noticesStore } from '@wordpress/notices';
-import { __, sprintf } from '@wordpress/i18n';
 import { store as blockEditorStore } from '../../../store';
 import { isFiltered } from '../../../store/utils';
-import { unlock } from '../../../lock-unlock';
 
 /**
  * Retrieves the block types inserter state.
@@ -35,10 +31,6 @@ const useBlockTypesState = ( rootClientId, onInsert, isQuick ) => {
 			),
 		[ rootClientId, options ]
 	);
-	const { getClosestAllowedInsertionPoint } = unlock(
-		useSelect( blockEditorStore )
-	);
-	const { createErrorNotice } = useDispatch( noticesStore );
 
 	const [ categories, collections ] = useSelect( ( select ) => {
 		const { getCategories, getCollections } = select( blocksStore );
@@ -57,26 +49,6 @@ const useBlockTypesState = ( rootClientId, onInsert, isQuick ) => {
 			},
 			shouldFocusBlock
 		) => {
-			const destinationClientId = getClosestAllowedInsertionPoint(
-				name,
-				rootClientId
-			);
-			if ( destinationClientId === null ) {
-				const title = getBlockType( name )?.title ?? name;
-				createErrorNotice(
-					sprintf(
-						/* translators: %s: block pattern title. */
-						__( 'Block "%s" can\'t be inserted.' ),
-						title
-					),
-					{
-						type: 'snackbar',
-						id: 'inserter-notice',
-					}
-				);
-				return;
-			}
-
 			const insertedBlock =
 				syncStatus === 'unsynced'
 					? parse( content, {
@@ -88,19 +60,9 @@ const useBlockTypesState = ( rootClientId, onInsert, isQuick ) => {
 							createBlocksFromInnerBlocksTemplate( innerBlocks ),
 							innerContent
 						);
-			onInsert(
-				insertedBlock,
-				undefined,
-				shouldFocusBlock,
-				destinationClientId
-			);
+			onInsert( insertedBlock, undefined, shouldFocusBlock );
 		},
-		[
-			getClosestAllowedInsertionPoint,
-			rootClientId,
-			onInsert,
-			createErrorNotice,
-		]
+		[ onInsert ]
 	);
 
 	return [ items, categories, collections, onSelectItem ];